Move config under service:runtype command
